Add files via upload
Understanding social and gender and studying information about students is a vital tool that can be used to understand what sort of grades a student is obtaining. Working as a teacher at a small secondary school, Ruth gave their students a survey to complete and tasked me to analyze the information and attempt to predict a student’s final grade.